backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.0.1
supabase==2.15.0
openai==1.59.7
google-genai==1.62.0
feedparser==6.0.11
requests==2.32.3
pydantic==2.10.6
